Move lambda to package
